pyproject.toml
[project]
name = "lottogogo"
version = "0.1.0"
description = "LottoGoGo probability engine sprint 1 implementation"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2,<3.0",
  "scipy>=1.14,<2.0",
  "pydantic>=2.0,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "mypy>=1.0,<2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
